spec2pix.py can now be used to generate the model image from the extracted spectra and psf fits files.  It now requires the HDU of the spectra to be specified (--hdu) since toy input spectra are in HDU 0, while the deconvolved output spectra are in HDU 3.
Added --noise option to spec2pix.py to add fake noise.  Default
is to not add noise (previous default was to add noise).

spec2pix.py does *not* propagate the extracted spectra covariance
into a pixel covariance.  If it isn't adding noise (--noise) it
just writes a zeros array to HDU 1 for the image ivar.

Updated test_data_model.sh and plot_test_results to use this
model image rather than creating it on the fly when plotting.